Add baseline results comparison
This commit adds baseline comparison logic to the
benchmark_regression_test workflow. The baseline artifact is uploaded
via the benchmark_baseline workflow once every 20 days. The baseline
artifact comparison will be skipped if the baseline result artifact is
not available for download, this is to handle the first 20 days after
merge.

Signed-off-by: Arjun Raja Yogidas <arjunry@amazon.com>