Update check regression script
This commit updates the regression check script to skip the initial
value in all BenchmarkTimes array of the benchmark results json file to
calculate a new p90. We use this new p90 to identify regression, this
change was made to combat the skewed p90 metrics we were seeing due to
the slow starts of the benchmark pull times which were affecting the
overall regression comparison. Skipping the first value allows us to
have a more uniform comparison, remove github environment noise and we'd be able to identify true regression in our code

Signed-off-by: Arjun Raja Yogidas <arjunry@amazon.com>